requirements.txt
django-grappelli==2.4.10